+++ test/lib/noseplugin.py
